+++ test/publications.tests.js
@@ -87,6 +87,26 @@ describe('Publications', function() {
         })
     })
 
+    it('should report deprecated publications', function(done) {
+        createBroker({
+            vhosts: vhosts,
+            publications: {
+                p1: {
+                    exchange: 'e1',
+                    deprecated: true
+                }
+            }
+        }, function(err, broker) {
+            assert.ifError(err)
+            broker.publish('p1', 'test message', function(err, publication) {
+                assert.ifError(err)
+                publication.on('success', function(messageId) {
+                    amqputils.assertMessage('q1', namespace, 'test message', done)
+                })
+            })
+        })
+    })
+
     it('should publish text messages to normal exchanges', function(done) {
         createBroker({
             vhosts: vhosts,